feat: atomizing classes
